pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "webserver"
version = "0.2.1"
authors = [
  { name="Nikolaos Chatzikonstantinou", email="nchatz314@gmail.com" },
]
description = "The Communal Growth webserver"
readme = "README.txt"
license = {file = "COPYING"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
]
dependencies = [
    "dateparser",
    "defusedxml",
    "lark",
    "litestar[standard, sqlalchemy]",
    "nameparser",
    "psycopg",
    "python-daemon",
    "requests",
    "watchdog",
]

[project.optional-dependencies]
tests = [
    "pytest",
    "pytest-mock",
    "pytest-xdist",
]
dev = [
    "black"
]

[project.scripts]
maildirdaemon = "app.maildirdaemon:main"

[project.urls]
Homepage = "https://github.com/communalgrowth/webserver"
Issues = "https://github.com/communalgrowth/webserver/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -m 'not (test_slow or test_online)'"
testpaths = [
    "tests",
]
markers = [
    "test_online: test requires internet connection.",
    "test_slow: test is slow."
]
